improved error handling

+++ main.py
@@ -27,9 +27,12 @@ def main(input: str) -> bool:
     # Initialize services and configuration
     transcriber = config.TRANSCRIBER
     assemblyai_token = config.ASSEMBLYAI_API_KEY if transcriber == "assemblyai" else None
-    
+    delete_after_upload = config.DELETE_AFTER_UPLOAD
     model = config.PREFERRED_MODELS.get(transcriber, None)
-    youtube_service = get_youtube_service() # Initialize YouTube API service, comment this line if you are not uploading to YouTube
+
+    # You need to have the client_secret.json file in the same directory as this script.
+    # If you don't, you can download it from the Google Cloud Console. An invalid client_secret.json will cause an error.
+    youtube_service = get_youtube_service()
 
     # Download Video
     downloaded_path = download_video_if_needed(input)
@@ -57,30 +60,42 @@ def main(input: str) -> bool:
     # Get list of already uploaded videos
     uploaded_videos = get_uploaded_videos()
 
-    # Upload to YouTube
+    # Youtube service check
+    if not youtube_service:
+        logging.error("YouTube service object is None. Exiting upload.")
+        return True
+
+    # Main loop to upload videos
+
     for idx, scene in enumerate(final_videos):
         if scene in uploaded_videos:
             logging.info(f"Video {scene} already uploaded. Skipping...")
-            os.remove(scene)
+            if delete_after_upload:
+                try:
+                    os.remove(scene)
+                    logging.info(f"Scene {idx+1} deleted. Path: {scene}")
+                except Exception as e:
+                    logging.error(f"Error deleting scene {idx+1}: {e}")
             continue
         success, limit = upload_scene(scene_path=scene, idx=idx, dir_name=target, youtube_service=youtube_service)
         if not success:
             log_failed_upload(scene)
             if limit:
-                logging.warning("Daily upload limit reached. Stopping uploads.")
                 break
             else:
                 continue
         log_uploaded_video(scene)
 
         #Delete local file after successful upload
-        try:
-            os.remove(scene)
-            logging.info(f"Scene {idx+1} deleted. Path: {scene}")
-        except Exception as e:
-            logging.error(f"Error deleting scene {idx+1}: {e}")
+        if delete_after_upload:
+            try:
+                os.remove(scene)
+                logging.info(f"Scene {idx+1} deleted. Path: {scene}")
+            except Exception as e:
+                logging.error(f"Error deleting scene {idx+1}: {e}")
         if idx < len(final_videos) - 1:
             sleep(10) # Delay before the next upload.
+
     return True
 
 def main_upload_only(scenes_directory: str):
@@ -93,37 +108,51 @@ def main_upload_only(scenes_directory: str):
     if not os.path.exists(scenes_directory):
         logging.error("Scenes directory does not exist. Exiting.")
         return False
-    youtube_service = get_youtube_service() # Initialize YouTube API service
-    uploaded_videos = get_uploaded_videos() # Get list of already uploaded videos
+
+    # You need to have the client_secret.json file in the same directory as this script.
+    # If you don't, you can download it from the Google Cloud Console. An invalid client_secret.json will cause an error.
+    youtube_service = get_youtube_service()
+    if not youtube_service:
+        logging.error("YouTube service object is None. Exiting upload.")
+        return False
+    
+    uploaded_videos = get_uploaded_videos()
+    delete_after_upload = config.DELETE_AFTER_UPLOAD
+
     target = os.path.dirname(scenes_directory)
     video_list = sorted(os.listdir(scenes_directory))
     video_list = [video for video in video_list if video.endswith(".mp4")] #Include only .mp4 files
+
+    # Main loop to upload videos
     for idx, scene in enumerate(video_list):
         if scene == ".DS_Store":  # Skip .DS_Store
             continue
 
         scene_path = os.path.join(scenes_directory, scene)
         if scene_path in uploaded_videos:
-            logging.info(f"Video {scene_path} already uploaded. Skipping...")
-            os.remove(scene_path)
+            if delete_after_upload:
+                try:
+                    os.remove(scene)
+                    logging.info(f"Scene {idx+1} deleted. Path: {scene}")
+                except Exception as e:
+                    logging.error(f"Error deleting scene {idx+1}: {e}")
             continue
-
         success, limit = upload_scene(scene_path=scene_path, idx=idx, dir_name=target, youtube_service=youtube_service)
         if not success:
             log_failed_upload(scene_path)
             if limit:
-                logging.warning("Daily upload limit reached. Stopping uploads.")
                 break
             else:
                 continue
-        log_uploaded_video(scene_path) # Log the uploaded video
+        #log_uploaded_video(scene_path) # Log the uploaded video
 
         #Delete local file after successful upload
-        try:
-            os.remove(scene_path)
-            logging.info(f"Scene {idx+1} deleted. Path: {scene}")
-        except Exception as e:
-            logging.error(f"Error deleting scene {idx+1}: {e}")
+        if delete_after_upload:
+            try:
+                os.remove(scene)
+                logging.info(f"Scene {idx+1} deleted. Path: {scene}")
+            except Exception as e:
+                logging.error(f"Error deleting scene {idx+1}: {e}")
         if idx < len(video_list) - 1:
             sleep(10) # Delay before the next upload.
     return True
